🚧 [plot] matplotlib heatmap
